# specs/001-ai-practice-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise AI Intelligent Practice System API
  description: REST API for PPT presentation coaching and sales practice bot
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Local development server
...
paths:
...
  /presentations:
    get:
      tags: [presentations]
      summary: List all presentations
      operationId: listPresentations
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ready, processing, failed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of presentations
          content:
            application/json:
              schema:
                type: object
                properties:
                  presentations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Presentation'
                  total:
                    type: integer
                  page:
                    type: integer

    post:
      tags: [presentations]
      summary: Upload a new presentation
      operationId: uploadPresentation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, title]
              properties:
                file:
                  type: string
                  format: binary
                  description: PPTX file (max 50MB)
                title:
                  type: string
                  description: Presentation title
      responses:
        '201':
          description: Presentation uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Presentation'
        '400':
          $ref: '#/components/responses/BadRequest'
  /presentations/{presentation_id}:
    get:
      tags: [presentations]
      summary: Get presentation details
      operationId: getPresentation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PresentationId'
      responses:
        '200':
          description: Presentation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [presentations]
      summary: Delete a presentation
      operationId: deletePresentation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PresentationId'
      responses:
        '204':
          description: Presentation deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /presentations/{presentation_id}/pages:
    get:
      tags: [presentations]
      summary: Get presentation pages
      operationId: getPresentationPages
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PresentationId'
      responses:
        '200':
          description: List of pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Page'
  /presentations/{presentation_id}/pages/{page_number}/talking-points:
    get:
      tags: [presentations]
      summary: Get required talking points for a page
      operationId: getTalkingPoints
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - name: page_number
          in: path
          required: true
          schema:
            type: integer
    post:
      tags: [presentations]
      summary: Add required talking point
      operationId: addTalkingPoint
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - name: page_number
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description]
              properties:
                description:
                  type: string
                  example: "Must mention our proprietary technology"
      responses:
        '201':
          description: Talking point added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequiredTalkingPoint'
  /presentations/{presentation_id}/forbidden-words:
    get:
      tags: [presentations]
      summary: Get forbidden words for presentation
      operationId: getForbiddenWords
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PresentationId'
    post:
      tags: [presentations]
      summary: Add forbidden word
      operationId: addForbiddenWord
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PresentationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phrase, suggested_alternative]
              properties:
                phrase:
                  type: string
                  example: "I don't know"
                suggested_alternative:
                  type: string
                  example: "Let me verify that information"
                page_id:
                  type: string
                  format: uuid
                  description: If specified, applies only to this page
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PresentationId:
      name: presentation_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Presentation:
      type: object
      properties:
        presentation_id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [ready, processing, failed]
        upload_date:
          type: string
          format: date-time
        total_pages:
          type: integer
    PresentationDetail:
      allOf:
        - $ref: '#/components/schemas/Presentation'
        - type: object
          properties:
            file_url:
              type: string
            version_number:
              type: integer
            pages:
              type: array
              items:
                $ref: '#/components/schemas/Page'
    Page:
      type: object
      properties:
        page_id:
          type: string
          format: uuid
        page_number:
          type: integer
        ocr_extracted_text:
          type: string
        image_url:
          type: string
        needs_manual_review:
          type: boolean
        talking_points:
          type: array
          items:
            $ref: '#/components/schemas/RequiredTalkingPoint'
    RequiredTalkingPoint:
      type: object
      properties:
        point_id:
          type: string
          format: uuid
        description:
          type: string
        is_ai_generated:
          type: boolean
        confirmed_by_admin:
          type: boolean
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: object
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Presentation not found"